Move predicates related to GRM to `resourcemanager` folder
