+++ pkg/resourcemanager/predicate/has_finalizer.go
